Make the template optional
Allows the user to be able to pass `false` to the grunt task to skip
template compilation.  This is useful for people who are looking to
compile markdown into partials by allowing them to skip the step of
creating an empty template like `<%= content %>`.

Does NOT change the default of using the included template.